Update INSTALL.md
Updated for the past two or three versions of Debian which do not include the `fastcgi` module and instead include the `fcgid` module which is also in the main repository for those that mind nonfree repositories. This change also requires removing the old `fastcgi` module directives which are no longer used and renaming the scripts themselves to end with `.fcgi`. The web app otherwise works out of the box.